Improve efficiency/correctness of event generation
Collect events to emit into two lists of Runnable, which can then
be executed later: one list for left and deleted events (called with
the old entity state) and one for created, entered and updated events
(called with the current entity state).

Make sure that for updated events, the list of changed FieldPaths is
sorted correctly.