Make entity parsing great again.
Version 2.5 suffered from buggy and slow entity parsing code.
Most of the problems stemmed from holding multiple client frames.

After being unable to sleep and thinking about this, I came to the
conclusion that the current mechanism was not needed, and that
one could still use only a single entity state, if taking special
cases, like double creates, double deletes (...) into account.

This is the result. It is as fast as 2.4, but can consume live data.